메뉴관리, 주문관리 추가

+++ src/main/java/com/example/weeklyquiz4/Repository/StoreRepository.java
@@ -3,5 +3,5 @@ package com.example.weeklyquiz4.repository;
 import com.example.weeklyquiz4.entity.StoreEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<StoreEntity, Integer> {
+public interface StoreRepository extends JpaRepository<StoreEntity, Long> {
 }
